Split out HTTPS only logic into another variable
There was an issue with WEB_REVERSE_PROXIED being on making it not possible to turn off HTTPS only logic, also the implicitness of WEB_HTTP=auto/true handling was confusing people.

In addition the separation makes it possible to fully turn off the http port.

The functionality over WEB_HTTP=auto/true is now deprecated but continues to function as before for BC (true turning off HTTPs redirect)