Replace mysql root DBA operations with optional admin user
It's not necessary for the application image to grant mysql permissions, as the mysql image does so, so this is deprecated.

However there are times when privilege-separation is useful, where the application user has less privileges than an admin user, so using an admin user is optional.

There is a BC break here where if a project relied on the default mysql root password, they wouldn't get the BC grant, and the application mysql user might not be set up previously, and so the app user credentials can't be used to do the mysql operations.

The workaround is for projects to define the DATABASE_ROOT_PASSWORD themselves, but better to add the environment variables to the mysql service instead to handle grants.