Make the container's plan.sh more accessible to projects
This allows applications to write the plan.sh in their project root path, and it get's copied in along with their project code.

Additionally cp-remote watch and local docker mounts will sync edits to it to the container, allowing realtime usage of the plan.sh functions and overrides without a new build or editing within the container.