Use all X-Forwarded-* headers by default
Some applications (e.g. Symfony) implicitly trust other X-Forwarded-* headers if they are sent from a trusted proxy. This means they can be manipulated by the user if not set or cleared.

Therefore a new "edge_headers" location setting, defaulting to true, can be used to say which trusted proxy is considered the edge server, and a new "hide_edge_headers" location setting to unset any earlier trusted proxy headers, when it's known there are none or not supplied.

X-Forwarded-For isn't just an edge header but a full chain, so it should be set whatever the case.

Sadly this wont cater to every reverse proxy situation, but is the best fit for most.