Backport Symfony app endpoint code to php images
Since PHP images are effectively doing the same thing, and other images aren't really related to Symfony but extending it just for this functionality, move it to the PHP images so it's more accessible.

Note there is a non-BC change here in the PHP image, in that the image no longer has the app endpoint enabled by default. This is because most