Remove magento1 dev package removal
Even if the php-memcache extension was already installed here, none of these packages were a result of it. If they are installed, they are installed in the parent images and it's the parent image's responsibility to remove unneeded packages.

(Also this means we don't need to pass PHP_VERSION as an arg)