Add the ability to schedule regeneration of dhparam in php-apache/php-nginx/nginx images
If a web server is not renewing it's dhparams regularly through new deployments or reloads (if using openssl's internally generated dhparam), then it's strength is reduced over time. A daily graceful reload of the webserver can re-read a regenerated dhparam file.

This requires START_CRON=true for the container to run the regeneration on it's schedule.